Agregado las validaciones

+++ src/main/java/com/teamcubation/AccionService/application/service/StockService.java
@@ -2,6 +2,9 @@ package com.teamcubation.AccionService.application.service;
 
 import com.teamcubation.AccionService.application.port.in.StockInPort;
 import com.teamcubation.AccionService.application.port.out.StockRepositoryPort;
+import com.teamcubation.AccionService.domain.exception.DuplicateStockException;
+import com.teamcubation.AccionService.domain.exception.InvalidStockModelException;
+import com.teamcubation.AccionService.domain.exception.StockNotFoundException;
 import com.teamcubation.AccionService.domain.model.Stock;
 import org.springframework.stereotype.Service;
 
@@ -18,12 +21,24 @@ public class StockService implements StockInPort {
     }
 
     @Override
-    public Stock create(Stock accion) {
+    public Stock create(Stock accion) throws DuplicateStockException, InvalidStockModelException {
+
+        if (isDuplicated(accion.getName())) {
+            throw new DuplicateStockException("Stock with that name already exists");
+        }
+
+        if (isInvalid(accion)) {
+            throw new InvalidStockModelException("Invalid stock data");
+        }
+
         return this.accionRepositoryPort.create(accion);
     }
 
     @Override
-    public Optional<Stock> findById(Long id) {
+    public Optional<Stock> findById(Long id) throws StockNotFoundException {
+        if (isNotFound(id)) {
+            throw new StockNotFoundException("Stock not found");
+        }
         return this.accionRepositoryPort.findById(id);
     }
 
@@ -33,12 +48,38 @@ public class StockService implements StockInPort {
     }
 
     @Override
-    public Stock update(Stock accion) {
+    public Stock update(Stock accion) throws InvalidStockModelException, StockNotFoundException {
+        if (isInvalid(accion)) {
+            throw new InvalidStockModelException("Invalid stock data");
+        }
+        if (isNotFound(accion.getId())) {
+            throw new StockNotFoundException("Stock not found");
+        }
         return this.accionRepositoryPort.update(accion);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws StockNotFoundException {
+        if (isNotFound(id)) {
+            throw new StockNotFoundException("Stock not found");
+        }
         this.accionRepositoryPort.deleteById(id);
+    }
+
+    //validations
+
+    private boolean isNotFound(Long id) {
+        return this.accionRepositoryPort.findById(id).isEmpty();
+    }
+
+    private boolean isDuplicated(String name) {
+        return this.accionRepositoryPort
+                .getAll()
+                .stream()
+                .anyMatch(accion -> accion.getName().equals(name));
+    }
+
+    private boolean isInvalid(Stock accion) {
+        return accion.getName() == null || accion.getPrice() < 0 || accion.getDividend() < 0;
     }
 }
